oracle_vertica/requirements.txt
apache-airflow
cx_Oracle
vertica-python
sqlalchemy
